[refact] : data 모듈화
