+++ app/build.gradle.kts
@@ -54,6 +54,7 @@ android {
 
 dependencies {
     implementation(project(":domain"))
+    implementation(project(":data-remote"))
 
     // coil
     implementation(libs.coil)
@@ -80,6 +81,8 @@ dependencies {
     implementation(libs.fragment.ktx)
     // dagger hilt
     implementation(libs.dagger.hilt)
+    implementation(project(":data-local"))
+    implementation(project(":data"))
     kapt(libs.dagger.hilt.compiler)
     // 기초 androidx 라이브러리 ("core-ktx", "constraintlayout", "appcompat", "activity")
     implementation(libs.bundles.androidx)
