--- app/src/main/java/com/sopt/now/data/api/ApiKeyStorage.kt
@@ -1,9 +1,0 @@
-package com.sopt.now.data.api
-
-object ApiKeyStorage {
-    const val MEMBER = "member"
-    const val JOIN = "join"
-    const val INFO = "info"
-    const val LOGIN = "login"
-    const val PASSWORD = "password"
-}
