--- app/src/main/java/com/sopt/now/data/repositoryimpl/extensions/ServerValidHttpException.kt
@@ -1,5 +1,0 @@
-package com.sopt.now.data.repositoryimpl.extensions
-
-data class ServerValidHttpException(
-    val errorMessage: String,
-) : Exception(errorMessage)
