+++ data-local/src/main/java/com/sopt/now/data_local/datasource/SharedPreferenceDataSource.kt
@@ -1,0 +1,11 @@
+package com.sopt.now.data_local.datasource
+
+import com.sopt.now.data_local.dto.UserDto
+
+interface SharedPreferenceDataSource {
+    var memberId: Int
+    fun saveUserInfo(userDto: UserDto?)
+    fun getUserInfo(): UserDto
+
+    fun clear()
+}
