+++ data-local/src/main/java/com/sopt/now/data_local/dto/UserDto.kt
@@ -1,0 +1,11 @@
+package com.sopt.now.data_local.dto
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class UserDto(
+    val id: String,
+    val password: String,
+    val nickName: String,
+    val phone: String
+)
