+++ data-remote/src/main/java/com/sopt/now/data_remote/api/ApiKeyStorage.kt
@@ -1,0 +1,8 @@
+package com.sopt.now.data_remote.api
+
+object ApiKeyStorage {
+    const val MEMBER = "member"
+    const val JOIN = "join"
+    const val INFO = "info"
+    const val LOGIN = "login"
+}
