+++ data-remote/src/main/java/com/sopt/now/data_remote/datasource/ReqresDataSource.kt
@@ -1,0 +1,7 @@
+package com.sopt.now.data_remote.datasource
+
+import com.sopt.now.data_remote.dto.response.ResponseReqresDto
+
+interface ReqresDataSource {
+    suspend fun getReqresList(page: Int): ResponseReqresDto
+}
