+++ data-remote/src/main/java/com/sopt/now/data_remote/dto/response/BaseResponse.kt
@@ -1,0 +1,14 @@
+package com.sopt.now.data_remote.dto.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class BaseResponse<T>(
+    @SerialName("code")
+    val code: Int,
+    @SerialName("message")
+    val message: String,
+    @SerialName("data")
+    val data: T? = null
+)
