+++ data-remote/src/main/java/com/sopt/now/data_remote/dto/response/ResponseMemberInfo.kt
@@ -1,0 +1,14 @@
+package com.sopt.now.data_remote.dto.response
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ResponseMemberInfo(
+    @SerialName("authenticationId")
+    val authenticationId: String,
+    @SerialName("nickname")
+    val nickname: String,
+    @SerialName("phone")
+    val phone: String
+)
